pyproject.toml
[tool.poetry]
name = "jan-2024-django3"
version = "0.1.0"
description = ""
authors = ["tetianaklius <tetyanaklyus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
#django-cors-headers = "^4.4.0"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
